[feature] Load app by type and title
Previously the only way to get the app was to get the forefront app offered by the OS, with this new feature we can identify an app among all the apps offered from the OS (currently on render) and pick one of them. This Feature is only avaiable for darwin platform. Make use of it on any other platform will return an error.